fix: refactor users routes from one file
